\name{dummy_CNVData10}
\alias{dummy_CNVData10}
\docType{data}
\title{
dummy_CNVData10
}
\description{
A R dataframe object with gene-centric copy number values. The first column, labeled as "sample", contains the sample IDs. The rest of the columns are the the values for the copy number values of the genes.
}
\format{
A data frame with 1225 observations for 10 samples. The first column contains the sample IDs and the rest of the columns are the observations.
}
\source{
TCGA Database (https://portal.gdc.cancer.gov/)
}
\examples{
data(dummy_CNVData10)
}
\keyword{datasets}
