\name{Dummy_RnaSeq_Count_Data}
\alias{Dummy_RnaSeq_Count_Data}
\docType{data}
\title{
Dummy_RnaSeq_Count_Data
}
\description{
A R dataframe object where the data is the RNA Seq count data for each patient and each gene in the study. The rows are the genes. The columns are the sample IDs for the patients.}
\format{
1125 observations of 374 variables
}
\source{
TCGA Database (https://portal.gdc.cancer.gov/)
}
\examples{
data(Dummy_RnaSeq_Count_Data)
## maybe str(Dummy_RnaSeq_Count_Data) ; plot(Dummy_RnaSeq_Count_Data) ...
}
\keyword{datasets}
