\name{dummy_RSeqData10_200}
\alias{dummy_RSeqData10_200}
\docType{data}
\title{
dummy_RSeqData10_200}
\description{
A R dataframe object. It contains the expression data fro the samples. The first column is labeled as "Sample" and contains the Sample IDs. The rest of the columns are the genes with expression values.
}
\format{
A data frame with 200 observations for 10 samples. The first column contains the sample IDs and the rest of the columns are the observations.
}
\source{
TCGA Database (https://portal.gdc.cancer.gov/)
}%%  ~~ reference to a publication or URL from which the data were obtained ~~

\examples{
data(dummy_RSeqData10_200)
## maybe str(dummy_RSeqData10_200) ; plot(dummy_RSeqData10_200) ...
}
\keyword{datasets}
