\name{dummy_TFData}
\alias{dummy_TFData}
\docType{data}
\title{
dummy_TFData}
\description{
A R dataframe object that contains gene and transcription factor(TF) interactions. The first column, labeled as "V1" contains the genes and the second column, labeled as "TF", contains the transcription factors.}
\format{
653 observations of 2 variables
}
\source{
TCGA Database (https://portal.gdc.cancer.gov/)
}%%  ~~ reference to a publication or URL from which the data were obtained ~~
\examples{
data(dummy_TFData)
## maybe str(dummy_TFData) ; plot(dummy_TFData) ...
}
\keyword{datasets}
