\name{dummy_mirnaData10}
\alias{dummy_mirnaData10}
\docType{data}
\title{
dummy_mirnaData10}
\description{
A R dataframe object. It contains the miRNA expression data for the samples. The first column is labeled as "sample" and contains the sample IDs. The rest of the columns are the miRNAs with expression values.
}
\format{
A data frame with 300 observations for 10 samples. The first column contains the sample IDs and the rest of the columns are the observations.
}
\source{
TCGA Database (https://portal.gdc.cancer.gov/)
}%%  ~~ reference to a publication or URL from which the data were obtained ~~
\examples{
data(dummy_mirnaData10)
## maybe str(dummy_mirnaData10) ; plot(dummy_mirnaData10) ...
}
\keyword{datasets}
