\name{dummy_TFData}
\alias{dummy_TFData}
\docType{data}
\title{
dummy_TFData}
\description{
A R dataframe object that contains gene and transcription factor(TF) interactions. The first column, labeled as "V1" contains the genes and the second column, labeled as "TF", contains the transcription factors.}
\format{
653 observations of 2 variables
}
\source{
TCGA Database (https://portal.gdc.cancer.gov/)
}
\examples{
data(miRDriverData)
dummy_TFData[1:3,]
}
\keyword{datasets}
