\name{Genes_coord}
\alias{Genes_coord}
\docType{data}
\title{Genes_coord}
\description{
A R dataframe object. It has the coordinates for all the genes in the human genome. There are four columns, such as the chromosome, the start position, the end position and the ENSEMBLE ID. The rows are labeled with the gene IDs.}
\format{
55746 observations of four variables
}
\source{
TCGA Database (https://portal.gdc.cancer.gov/)
}%%  ~~ reference to a publication or URL from which the data were obtained ~~

\examples{
data(Genes_coord)
## maybe str(Genes_coord) ; plot(Genes_coord) ...
}
\keyword{datasets}
