\name{getRegionWiseGistic}
\alias{getRegionWiseGistic}
\title{
produce regionwise Gistic files i.e. peak wise files
}
\description{
getRegionWiseGistic will produce regionwise Gistic files i.e. aberration peak wise files
}
\usage{
getRegionWiseGistic(gisticfile,gistic_bedfile)
}
\arguments{
    \item{gisticfile}{
a character string for the memory location of the gistic file or amplification/deletion peaks with patients file in .txt
} \item{gistic_bedfile}{
a logical value for whether or not one wants a gistic_bedfile
}
}\author{
Banabithi Bose
}\examples{
print("Produce regionwise Gistic files i.e. aberration peak wise files")
##The following code is a valid example of how to use the function but is commented as it
##takes a very long time to run
##gisticfile<-"~/mirDriverFold/EXAMPLE_DATASETS/all_lesions.conf_90.txt"
##getRegionWiseGistic(gisticfile=gisticfile,gistic_bedfile=TRUE)
}\value{
Returns nothing
}